Add Images for TESTING.md
I have added the images for the automated tests to this file. I have also updated the UATs so that they have links to the correctly-formatted PDFs that I made last night. I used this new technique I saw (http://bit.ly/23dB7DY) that allows us to link directly to a file in our GitHub repository, rather than posting it to Imgur and doing it that way. If you wish for me to change to that method, let me know and I will make that change. Your input is appreciated! :)